SI-4489 feat(notes): Вынес маппинг заметок в интерактор
